feat(example): fix example and add atomic-counters
fix example and add atomic-counters
