+++ GoExample/arrays/arrays.go
