+++ GoExample/closures/closures.go
