+++ GoExample/constant/constant.go
