+++ GoExample/embedding/embedding.go
