+++ GoExample/for/for.go
