+++ GoExample/functions/functions.go
