+++ GoExample/generics/generics.go
