+++ GoExample/interfaces/interfaces.go
