+++ GoExample/methods/methods.go
