+++ GoExample/multiple-return-values/multiple-return-values.go
