+++ GoExample/non-blocking-channel-operations/non-blocking-channel-operations.go
@@ -5,7 +5,8 @@ import (
 )
 
 // 常规的通过通道发送和接收数据是阻塞的
-// 然而可以使用带一个 default 子句的 select 来实现非阻塞 的发送、接收，甚至是非阻塞的多路 select
+// 然而可以使用带一个 default 子句的 select 来实现非阻塞 的发送、接收
+// 甚至是非阻塞的多路 select
 func main() {
 	messages := make(chan string)
 	signals := make(chan bool)
@@ -21,6 +22,9 @@ func main() {
 	}
 
 	// 一个非阻塞发送的实现方法
+	//  msg 不能被发送到 message 通道
+	//  因为这是 个无缓冲区通道，并且也没有接收者
+	//  因此， default 会执行
 	msg := "hi"
 	select {
 	case messages <- msg:
@@ -30,7 +34,7 @@ func main() {
 	}
 
 	// 可以在 default 前使用多个 case 子句来实现一个多路的非阻塞的选择器
-	// 这里试图在 messages和 signals 上同时使用非阻塞的接受操作
+	// 这里试图在 messages和 signals 上同时使用非阻塞的接收操作
 	select {
 	case msg := <-messages:
 		fmt.Println("received message", msg)
