+++ GoExample/pointers/pointers.go
