+++ GoExample/slices/slices.go
