+++ GoExample/structs/structs.go
