+++ GoExample/switch/switch.go
