+++ GoExample/timeouts/timeouts.go
@@ -5,10 +5,11 @@ import (
 	"time"
 )
 
-// 超时 对于一个连接外部资源，或者其它一些需要花费执行时间的操作的程序而言是很重要的
+// 超时 对于一个需要连接外部资源，或者有耗时较长的操作的程序而言是很重要的
 // 得益于通道和 select，在 Go中实现超时操作是简洁而优雅的
+
 func main() {
-	// 假如执行一个外部调用，并在 2 秒后通过通道 c1 返回它的执行结果
+	// 假如执行一个外部调用，并在 2 秒后，使用通道 c1 返回它的执行结果
 	c1 := make(chan string, 1)
 	go func() {
 		time.Sleep(time.Second * 2)
@@ -16,8 +17,8 @@ func main() {
 	}()
 
 	// 使用 select 实现一个超时操作
-	// res := <- c1 等待结果，<-Time.After 等待超时时间 1 秒后发送的值
-	// 由于 select 默认处理第一个已准备好的接收操作，如果这个操作超过了允许的 1 秒的话，将会执行超时 case
+	// res := <- c1 等待结果，<-time.After 等待超时 (1 秒钟）后发送的值
+	// 由于 select 默认处理第一个已准备好的接收操作，因此如果操作耗时超过了允许的 1 秒的话，将会执行超时 case
 	select {
 	case res := <-c1:
 		fmt.Println(res)
@@ -37,6 +38,5 @@ func main() {
 	case <-time.After(time.Second * 3):
 		fmt.Println("timeout 2")
 	}
-	// 使用这个 select 超时方式，需要使用通道传递结果
-	// 这对于一般情况是个好的方式，因为其他重要的 Go 特性是基于通道和select 的
+	// 运行这个程序，首先显示运行超时的操作，然后是成功接收的
 }
