+++ GoExample/variables/variables.go
