+++ GoExample/variadic-functions/variadic-functions.go
